add deepCopy unit test
