+++ src/plugins/logger.js
@@ -1,5 +1,7 @@
 // Credits: borrowed code from fcomb/redux-logger
 
+import { deepCopy } from '../util'
+
 export default function createLogger ({
   collapsed = true,
   transformer = state => state,
@@ -49,40 +51,4 @@ function repeat (str, times) {
 
 function pad (num, maxLength) {
   return repeat('0', maxLength - num.toString().length) + num
-}
-
-function find (list, f) {
-  return list.filter(f)[0]
-}
-
-/**
- * Deep copy the given object considering circular structure.
- * This function caches all nested objects and its copies.
- * If it detects circular structure, use cached copy to avoid infinite loop.
- */
-function deepCopy (obj, cache = []) {
-  // just return if obj is immutable value
-  if (obj === null || typeof obj !== 'object') {
-    return obj
-  }
-
-  // if obj is hit, it is in circular structure
-  const hit = find(cache, c => c.original === obj)
-  if (hit) {
-    return hit.copy
-  }
-
-  const copy = Array.isArray(obj) ? [] : {}
-  // put the copy into cache at first
-  // because we want to refer it in recursive deepCopy
-  cache.push({
-    original: obj,
-    copy
-  })
-
-  Object.keys(obj).forEach(key => {
-    copy[key] = deepCopy(obj[key], cache)
-  })
-
-  return copy
 }
